requirements.txt
streamlit
pandas
chromadb
ollama
openpyxl
unstructured
gradio
fastembed
faiss-cpu
llama-index-vector-stores-faiss